Add experimental font spacing feature
This helps with the "Style 1" issue and sets line spacing based on the font you have configured there. Note that it currently operates on a set list of fonts and sizes. If a font is not in the list, it goes back go the old default and emits a warning.