Better split pages in Doblon to KBS conversion
If there are more than max_lines_per_page on a page after splitting only by empty lines and time gaps, split into multiple pages in such a way that the total number of pages is minimal, and each page has at most one more line than any other in a split set. E.g.

Group of 7 lines with max_lines_per_page of 6:
- creates two pages since 7 lines can't fit on one page
- makes the pages have 4 and 3 lines respectively to split as evenly as possible

Group of 16 lines with max_lines_per_page of 5:
- Creates 4 pages since 3*5=15 <= 16 <= 20=4*5
- The pages then have 5, 4, 4, 4 lines

Some of this logic should be generalized/moved so it can be used in an eventual .lrc converter or any other format that has timing info with no layout.