Initial scrolling feature support in kbp2ass
Adds a CLI option for applying effects, the first one being scrolling, and some scrolling-related options:

  --effects [{SCROLLING} ...], -e [{SCROLLING} ...]
                        Iterable (default: ())
  --scrolling-type {IN_LINE_START,OUT_LINE_END}, -c {IN_LINE_START,OUT_LINE_END}
                        AssScroll (default: IN_LINE_START)
  --scrolling-duration SCROLLING_DURATION, -d SCROLLING_DURATION
                        int (default: 500)
  --scrolling-offset SCROLLING_OFFSET, -g SCROLLING_OFFSET
                        int (default: -50)

If scrolling is enabled, any page with lines going off screen will cause that page to scroll through the lines.

If scrolling_type IN_LINE_START is set, the scrolling occurs just before the words on the current line start wiping. If OUT_LINE_END is specified, scrolling occurs just before the words on a line finish wiping.

The scrolling_duration option specifies how long it should take to move a line up one position.

The scrolling_offset option applies an offset to the start of scrolling, so for example, with -50, duration 500, and IN_LINE_START, scrolling will start 550ms before text on the current line starts wiping, and will complete 50ms before.

This code still needs a lot of work to handle edge cases, but it is usable if some manual editing of the resulting .ass file can be tolerated.