CLI help improvements (show better help, add --usage-all)
Embed help info into options dataclasses through the metadata "doc" attribute and use that in the CLI. Implemented for kbp2ass subcommand, will implement for the others after some of the code is consolidated (so it only needs to be done once).

Add --usage-all to show the usage for all the subcommands at once. May consider doing the same for help, but it seems like too much info to show someone at a time.